Plan 9 from Bell Labs Fourth Edition 2006-05-22

+++ sys/src/cmd/unix/drawterm/gui-x11/xmem.h
@@ -31,7 +31,7 @@ struct Xmem
 	int	dirty;
 	Rectangle dirtyr;
 	Rectangle r;
-	ulong pc;	/* who wrote into xi */
+	uintptr pc;	/* who wrote into xi */
 };
 
 extern	int		xtblbit;
